Added a GUI. Still have some bugs in it that need to be hashed out.
Speed is totally gone now too. When before it'd be almost instant even on rot13big.ook, now it's got many seconds of lag. Also true for squares. This would be more okay if the GUI actually printed out characters as they were made, but it seems to be buffering them away somewhere.